+++ src/index.ts
@@ -5,6 +5,7 @@ export {
   type MergeOptions,
   type PivotTableOptions,
   type ToCSVOptions,
+  type ToJSONOptions,
   type ValueCountsOptions,
 } from "./dataframe";
 export { GroupBy } from "./groupby";
@@ -12,12 +13,16 @@ export {
   concat,
   merge,
   parse_csv,
+  parse_json,
   pivot_table,
   read_csv,
   read_csv_sync,
+  read_json,
+  read_json_sync,
   to_csv,
   type ConcatOptions,
   type ReadCSVOptions,
+  type ReadJSONOptions,
 } from "./io";
 export { Series, type SeriesDType, type SeriesOptions } from "./series";
 export type {